* changes:
  fix(xilinx): resolve misra rule 16.3 violations
  fix(xilinx): resolve misra rule 2.5 violations
  fix(xilinx): resolve misra rule 4.6 violations
  fix(xilinx): resolve misra rule 12.2 violations
  fix(xilinx): resolve misra rule 10.1 violations
  fix(xilinx): resolve misra rule 8.13 violations
  fix(xilinx): resolve misra rule 4.5 violations
  fix(xilinx): resolve misra rule 16.4 violations

+++ plat/xilinx/common/include/pm_common.h
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2013-2018, Arm Limited and Contributors. All rights reserved.
- * Copyright (c) 2022-2024, Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (c) 2022-2025, Advanced Micro Devices, Inc. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -29,9 +29,9 @@
 #define RET_PAYLOAD_ARG_CNT	6U
 #define PAYLOAD_ARG_SIZE	4U	/* size in bytes */
 
-#define TZ_VERSION_MAJOR	1
-#define TZ_VERSION_MINOR	0
-#define TZ_VERSION		((TZ_VERSION_MAJOR << 16) | \
+#define TZ_VERSION_MAJOR	1U
+#define TZ_VERSION_MINOR	0U
+#define TZ_VERSION		(((uint32_t)TZ_VERSION_MAJOR << 16U) | \
 				 TZ_VERSION_MINOR)
 
 /**
